move events to native-bridge.js
